+++ kfinance/kfinance.py
@@ -1644,8 +1644,11 @@ class Client:
             for tool_cls in ALL_TOOLS:
                 tool = tool_cls(kfinance_client=self)  # type: ignore[call-arg]
                 if (
-                    tool.required_permission is None
-                    or tool.required_permission in self.kfinance_api_client.user_permissions
+                    tool.accepted_permissions is None
+                    # if one or more of the required permission for a tool is a permission the user has
+                    or tool.accepted_permissions.intersection(
+                        self.kfinance_api_client.user_permissions
+                    )
                 ):
                     self._tools.append(tool)